requirements.txt
APScheduler==3.11.0
brotli==1.1.0
ConfigParser==7.1.0
cryptography==44.0.1
Cython==3.0.12
dl==0.1.0
docutils==0.21.2
Flask==3.1.0
flask_cors==5.0.1
folium==0.19.4
geopy==2.4.1
HTMLParser==0.0.2
importlib_metadata==8.6.1
ipaddr==2.2.0
keyring==25.6.0
lockfile==0.12.2
lxml==5.3.1
matplotlib==3.10.0
mypy_extensions==1.0.0
mysql_connector_repackaged==0.3.1
networkx==3.4.2
numpy==2.2.3
ordereddict==1.1
pandas==2.2.3
protobuf==5.29.3
pyOpenSSL==25.0.0
pytz==2025.1
seaborn==0.13.2
SQLAlchemy==2.0.38
thread==2.0.5
typing_extensions==4.12.2
wincertstore==0.2.1
xmlrpclib==1.0.1
zipp==3.21.0
